Fix prioritized 0.10 bugs (#250 #259 #253 #271 #216 #261 #260 #224)
serialize.rs:
- #250: remove the `if false && do_subset` guard that hard-disabled font
  subsetting; subsetting now runs when do_subset && requires_subsetting
  && embedding_mode == Subset. (The production subset path uses
  allsorts::subset directly and falls back to the full font on error,
  so the masked source-bytes panic does not recur here.)
- #259: rectangle_to_stream_ops now honors rectangle.mode — emits
  S/f/B/B* (and W..n for Clip) per PaintMode like polygon_to_stream_ops,
  instead of always `n` (which made rectangles invisible).
- #271: FontDescriptor Ascent/Descent/FontBBox are scaled by
  1000/units_per_em (glyph space), and CapHeight is the real OS/2
  sCapHeight (scaled), falling back to the scaled ascent instead of 0.
- #261: collect_used_glyphs_from_pages now always records the gid, even
  when the reverse char lookup fails and cid is None, so the font still
  gets embedded (the fallback char only affects ToUnicode).

graphics.rs:
- #253: Rect::to_array emits Real(f32), not `.round() as i64`, so
  media/trim/crop boxes are no longer distorted.

deserialize.rs:
- #216: get_dict_or_resolve_ref accepts Object::Stream (returns
  &stream.dict) and resolves references that point at stream objects,
  so XObject stream refs parse.

font.rs:
- #260: no-text_layout ParsedFont::from_bytes stub takes index: usize
  (matches azul's ParsedFont::from_bytes signature).

lib.rs:
- #224: with_pages now REPLACES self.pages (matching its doc);
  append_document remains for merging.

Adds focused unit tests for #253, #259, #224 (all pass). The pre-existing
font::test::subset_test failure is from azul-layout 0.0.9's ParsedFont::subset
source-bytes change and is unrelated to these fixes.

+++ src/font.rs
@@ -64,10 +64,10 @@ impl Default for PdfFontMetricsStub {
 
 #[cfg(not(feature = "text_layout"))]
 impl ParsedFont {
-    pub fn from_bytes(bytes: &[u8], index: u32, _warnings: &mut Vec<String>) -> Option<Self> {
+    pub fn from_bytes(bytes: &[u8], index: usize, _warnings: &mut Vec<String>) -> Option<Self> {
         Some(ParsedFont {
             original_bytes: bytes.to_vec(),
-            font_index: index,
+            font_index: index as u32,
             font_name: None,
             codepoint_to_glyph: BTreeMap::new(),
             glyph_widths: BTreeMap::new(),
